filtrerbar sökning på spelare klar

+++ src/main/java/com/example/databasteamecho/model/Player.java
@@ -103,7 +103,7 @@ public class Player {
         return phonenumber;
     }
 
-    public void setPhonenumber(int phonenumber) {
+    public void setPhonenumber(long phonenumber) {
         this.phonenumber = phonenumber;
     }
